Module anoma_vm_env::filter_prelude::storage

Storage types

Modules

testing
Helpers for testing with storage types.

Structs

BlockHash
Hash of a block as fixed-size byte array

BlockHeight
Height of a block, i.e. the level.

Key
A storage key is made of storage key segments DbKeySeg, separated by KEY_SEGMENT_SEPARATOR.

Enums

DbKeySeg
A storage key segment

Error

Constants

BLOCK_HASH_LENGTH
The length of the block’s hash string

CHAIN_ID_LENGTH
The length of chain ID string

KEY_SEGMENT_SEPARATOR
The separator of storage key segments

RESERVED_ADDRESS_PREFIX
The reserved storage key prefix for addresses

RESERVED_VP_KEY
The reserved storage key for validity predicates

VP_KEY_PREFIX
The reserved storage key prefix for validity predicates

Traits

KeySeg
Represents a segment in a path that may be used as a database key

Type Definitions

Result
Result for functions that may fail