Struct anoma_vm_env::filter_prelude::time::Duration

pub struct Duration { /* fields omitted */ }

ISO 8601 time duration with nanosecond precision. This also allows for the negative duration; see individual methods for details.

Implementations

impl Duration

pub fn weeks(weeks: i64) -> Duration

Makes a new Duration with given number of weeks. Equivalent to Duration::seconds(weeks * 7 * 24 * 60 * 60) with overflow checks. Panics when the duration is out of bounds.

pub fn days(days: i64) -> Duration

Makes a new Duration with given number of days. Equivalent to Duration::seconds(days * 24 * 60 * 60) with overflow checks. Panics when the duration is out of bounds.

pub fn hours(hours: i64) -> Duration

Makes a new Duration with given number of hours. Equivalent to Duration::seconds(hours * 60 * 60) with overflow checks. Panics when the duration is out of bounds.

pub fn minutes(minutes: i64) -> Duration

Makes a new Duration with given number of minutes. Equivalent to Duration::seconds(minutes * 60) with overflow checks. Panics when the duration is out of bounds.

pub fn seconds(seconds: i64) -> Duration

Makes a new Duration with given number of seconds. Panics when the duration is more than i64::MAX milliseconds or less than i64::MIN milliseconds.

pub fn milliseconds(milliseconds: i64) -> Duration

Makes a new Duration with given number of milliseconds.

pub fn microseconds(microseconds: i64) -> Duration

Makes a new Duration with given number of microseconds.

pub fn nanoseconds(nanos: i64) -> Duration

Makes a new Duration with given number of nanoseconds.

pub fn span<F>(f: F) -> Duration where
F: FnOnce(),

Runs a closure, returning the duration of time it took to run the closure.

pub fn num_weeks(&self) -> i64

Returns the total number of whole weeks in the duration.

pub fn num_days(&self) -> i64

Returns the total number of whole days in the duration.

pub fn num_hours(&self) -> i64

Returns the total number of whole hours in the duration.

pub fn num_minutes(&self) -> i64

Returns the total number of whole minutes in the duration.

pub fn num_seconds(&self) -> i64

Returns the total number of whole seconds in the duration.

pub fn num_milliseconds(&self) -> i64

Returns the total number of whole milliseconds in the duration,

pub fn num_microseconds(&self) -> Option<i64>

Returns the total number of whole microseconds in the duration, or None on overflow (exceeding 263 microseconds in either direction).

pub fn num_nanoseconds(&self) -> Option<i64>

Returns the total number of whole nanoseconds in the duration, or None on overflow (exceeding 263 nanoseconds in either direction).

pub fn checked_add(&self, rhs: &Duration) -> Option<Duration>

Add two durations, returning None if overflow occurred.

pub fn checked_sub(&self, rhs: &Duration) -> Option<Duration>

Subtract two durations, returning None if overflow occurred.

pub fn min_value() -> Duration

The minimum possible Duration: i64::MIN milliseconds.

pub fn max_value() -> Duration

The maximum possible Duration: i64::MAX milliseconds.

pub fn zero() -> Duration

A duration where the stored seconds and nanoseconds are equal to zero.

pub fn is_zero(&self) -> bool

Returns true if the duration equals Duration::zero().

pub fn from_std(duration: Duration) -> Result<Duration, OutOfRangeError>

Creates a time::Duration object from std::time::Duration

This function errors when original duration is larger than the maximum value supported for this type.

pub fn to_std(&self) -> Result<Duration, OutOfRangeError>

Creates a std::time::Duration object from time::Duration

This function errors when duration is less than zero. As standard library implementation is limited to non-negative values.