Module anoma_vm_env::imports::matchmaker

Matchmaker environment imports

Traits

BorshDeserialize
A data-structure that can be de-serialized from binary format by NBOR.

BorshSerialize
A data-structure that can be serialized into binary format by NBOR.

Functions

log_string
Log a string. The message will be printed at the tracing::Level::Info.

remove_intents
Remove the intents from the matchmaker intent mempool, to call when they are fulfilled or outdated.

send_match
Send a transaction with the tx_data and the tx_code to the ledger given in matchmaker parameters (--tx-code-path and --ledger-address).

update_data
Update the matchmaker state. This state will be pass on the next run of the matchmaker.

Derive Macros

BorshDeserialize

BorshSerialize