Trait anoma_vm_env::imports::tx::BorshSerialize

pub trait BorshSerialize {
    pub fn serialize<W>(&self, writer: &mut W) -> Result<(), Error>
    where
        W: Write;

    pub fn try_to_vec(&self) -> Result<Vec<u8, Global>, Error> { ... }

    pub unsafe fn is_u8() -> bool { ... }
}

A data-structure that can be serialized into binary format by NBOR.

Required methods

pub fn serialize<W>(&self, writer: &mut W) -> Result<(), Error> where
W: Write,

Loading content...
Provided methods

pub fn try_to_vec(&self) -> Result<Vec<u8, Global>, Error>

Serialize this instance into a vector of bytes.

pub unsafe fn is_u8() -> bool

Whether Self is u8. NOTE: Vec<u8> is the most common use-case for serialization and deserialization, it’s worth handling it as a special case to improve performance. It’s a workaround for specific Vec<u8> implementation versus generic Vec<T> implementation. See https://github.com/rust-lang/rfcs/pull/1210 for details.

It’s marked unsafe, because if the type is not u8 it leads to UB. See related issues:

https://github.com/near/borsh-rs/issues/17

https://github.com/near/borsh-rs/issues/18

Loading content...