Trait anoma_vm_env::imports::vp::BorshDeserialize

pub trait BorshDeserialize {
    fn deserialize(buf: &mut &[u8]) -> Result<Self, Error>;

    fn try_from_slice(v: &[u8]) -> Result<Self, Error> { ... }
}

A data-structure that can be de-serialized from binary format by NBOR.

Required methods

fn deserialize(buf: &mut &[u8]) -> Result<Self, Error>

Deserializes this instance from a given slice of bytes. Updates the buffer to point at the remaining bytes.

Provided methods

fn try_from_slice(v: &[u8]) -> Result<Self, Error>

Deserialize this instance from a slice of bytes.