Trait anoma_vm_env::key::vp::TryFromRef

pub trait TryFromRef<T> where
    T: ?Sized, {
    type Error;
    fn try_from_ref(value: &T) -> Result<Self, Self::Error>;
}

Simple and safe type conversions that may fail in a controlled way under some circumstances.

Associated Types

type Error

The type returned in the event of a conversion error.

Required methods

fn try_from_ref(value: &T) -> Result<Self, Self::Error>

Performs the conversion.