Module anoma_vm_env::matchmaker_prelude::transaction

Types that are used in transactions.

Modules

decrypted

txs that contain decrypted payloads or assertions of non-decryptability

pos

Types used for PoS system transactions

tx_types

Module that includes helper functions for classifying different types of transactions that the ledger must support as well as conversion functions between them.

wrapper

wrapper txs with encrypted payloads

Structs

Fee

A fee is an amount of a specified token

GasLimit

Gas limits must be multiples of GAS_LIMIT_RESOLUTION This is done to minimize the amount of information leak from a wrapper tx. The larger the GAS_LIMIT_RESOLUTION, the less info leaked.

Hash

A hash, typically a sha-2 hash of a tx

InitAccount

A tx data type to initialize a new established account

InitValidator

A tx data type to initialize a new validator account and its staking reward account.

UpdateVp

A tx data type to update an account’s validity predicate

WrapperTx

A transaction with an encrypted payload as well as some non-encrypted metadata for inclusion and / or verification purposes

Enums

DecryptedTx

Holds the result of attempting to decrypt a transaction and the data necessary for other validators to verify

TxType

Struct that classifies that kind of Tx based on the contents of its data.

WrapperTxErr

Errors relating to decrypting a wrapper tx and its encrypted payload from a Tx type

Traits

AffineCurve

Affine representation of an elliptic curve point guaranteed to be in the correct prime order subgroup.

PairingEngine

Functions

hash_tx

Get the hash of a transaction

process_tx

Determines the type of the input Tx

verify_decrypted_correctly

Verify that if the encrypted payload was marked “undecryptable”, we should not be able to decrypt it

Type Definitions

EllipticCurve