Module anoma_vm_env::proof_of_stake::epoched

Epoched and EpochedDelta are structures for data that is set for future epochs at a given EpochOffset.

Structs

Epoched
Data that may have values set for future epochs, up to an epoch at offset as set via the Offset type parameter.

EpochedDelta
Data that may have delta values (a difference from the predecessor epoch) set for future epochs, up to an epoch at offset as set via the Offset type parameter.

OffsetPipelineLen
Offset at pipeline length.

OffsetUnboundingLen
Offset at unbonding length.

Enums

DynEpochOffset
Offset length dynamic choice.

Traits

EpochOffset
Which offset should be used to set data. The value is read from PosParams.