Trait anoma_vm_env::proof_of_stake::vp::anoma_proof_of_stake::btree_set::BTreeSetShims

pub trait BTreeSetShims<T> {
    fn first_shim(&self) -> Option<&T>;
    fn last_shim(&self) -> Option<&T>;
    fn pop_first_shim(&mut self) -> Option<T>;
    fn pop_last_shim(&mut self) -> Option<T>;
}

This trait adds shims for BTreeSet methods that not yet stable. They have the same behavior as their nightly counterparts, but additionally require Clone bound on the element type (for pop_first and pop_last).

Required methods

fn first_shim(&self) -> Option<&T>

Returns a reference to the first value in the set, if any. This value is always the minimum of all values in the set.

fn last_shim(&self) -> Option<&T>

Returns a reference to the last value in the set, if any. This value is always the maximum of all values in the set.

fn pop_first_shim(&mut self) -> Option<T>

Removes the first value from the set and returns it, if any. The first value is always the minimum value in the set.

fn pop_last_shim(&mut self) -> Option<T>

Removes the last value from the set and returns it, if any. The last value is always the maximum value in the set.