Trait anoma_vm_env::proof_of_stake::vp::anoma_proof_of_stake::PosBase

pub trait PosBase {
    type Address: 'static + Display + Debug + Clone + PartialEq<Self::Address> + Eq + PartialOrd<Self::Address> + Ord + Hash + BorshDeserialize + BorshSerialize;
    type TokenAmount: 'static + Display + Debug + Default + Clone + Copy + Add<Self::TokenAmount> + AddAssign<Self::TokenAmount> + Sub<Self::TokenAmount> + PartialOrd<Self::TokenAmount> + Into<u64> + From<u64> + Into<Self::TokenChange> + SubAssign<Self::TokenAmount> + BorshDeserialize + BorshSerialize;
    type TokenChange: 'static + Display + Debug + Default + Clone + Copy + PartialOrd<Self::TokenChange> + Add<Self::TokenChange> + Sub<Self::TokenChange> + From<Self::TokenAmount> + From<i128> + Into<i128> + Neg + BorshDeserialize + BorshSerialize;
    type PublicKey: 'static + Debug + Clone + BorshDeserialize + BorshSerialize;

    const POS_ADDRESS: Self::Address;
    const POS_SLASH_POOL_ADDRESS: Self::Address;

    fn staking_token_address() -> Self::Address;

    fn read_pos_params(&self) -> PosParams;

    fn read_validator_address_raw_hash(
        &self, 
        raw_hash: impl AsRef<str>
    ) -> Option<Self::Address>;

    fn read_validator_consensus_key(
        &self, 
        key: &Self::Address
    ) -> Option<Epoched<Self::PublicKey, OffsetPipelineLen>>;

    fn read_validator_state(
        &self, 
        key: &Self::Address
    ) -> Option<Epoched<ValidatorState, OffsetPipelineLen>>;

    fn read_validator_total_deltas(
        &self, 
        key: &Self::Address
    ) -> Option<EpochedDelta<Self::TokenChange, OffsetUnboundingLen>>;

    fn read_validator_voting_power(
        &self, 
        key: &Self::Address
    ) -> Option<EpochedDelta<VotingPowerDelta, OffsetUnboundingLen>>;

    fn read_validator_slashes(&self, key: &Self::Address) -> Vec<Slash, Global>;

    fn read_validator_set(
        &self
    ) -> Epoched<ValidatorSet<Self::Address>, OffsetUnboundingLen>;

    fn read_total_voting_power(
        &self
    ) -> EpochedDelta<VotingPowerDelta, OffsetUnboundingLen>;

    fn write_pos_params(&mut self, params: &PosParams);

    fn write_validator_address_raw_hash(&mut self, address: &Self::Address);

    fn write_validator_staking_reward_address(
        &mut self, 
        key: &Self::Address, 
        value: &Self::Address
    );

    fn write_validator_consensus_key(
        &mut self, 
        key: &Self::Address, 
        value: &Epoched<Self::PublicKey, OffsetPipelineLen>
    );

    fn write_validator_state(
        &mut self, 
        key: &Self::Address, 
        value: &Epoched<ValidatorState, OffsetPipelineLen>
    );

    fn write_validator_total_deltas(
        &mut self, 
        key: &Self::Address, 
        value: &EpochedDelta<Self::TokenChange, OffsetUnboundingLen>
    );

    fn write_validator_voting_power(
        &mut self, 
        key: &Self::Address, 
        value: &EpochedDelta<VotingPowerDelta, OffsetUnboundingLen>
    );

    fn write_validator_slash(&mut self, validator: &Self::Address, value: Slash);

    fn write_bond(
        &mut self, 
        key: &BondId<Self::Address>, 
        value: &EpochedDelta<Bond<Self::TokenAmount>, OffsetPipelineLen>
    );

    fn write_validator_set(
        &mut self, 
        value: &Epoched<ValidatorSet<Self::Address>, OffsetUnboundingLen>
    );

    fn write_total_voting_power(
        &mut self, 
        value: &EpochedDelta<VotingPowerDelta, OffsetUnboundingLen>
    );

    fn init_staking_reward_account(
        &mut self, 
        address: &Self::Address, 
        pk: &Self::PublicKey
    );

    fn credit_tokens(
        &mut self, 
        token: &Self::Address, 
        target: &Self::Address, 
        amount: Self::TokenAmount
    );

    fn transfer(
        &mut self, 
        token: &Self::Address, 
        amount: Self::TokenAmount, 
        src: &Self::Address, 
        dest: &Self::Address
    );

    fn init_genesis<'a>(
        &mut self, 
        params: &'a PosParams, 
        validators: impl Iterator<Item = &'a GenesisValidator<Self::Address, Self::TokenAmount, Self::PublicKey>> + Clone + 'a, 
        current_epoch: impl Into<Epoch>
    ) -> Result<(), GenesisError> { ... }

    fn validator_set_update(
        &self, 
        current_epoch: impl Into<Epoch>, 
        f: impl FnMut(ValidatorSetUpdate<Self::PublicKey>)
    ) { ... }

    fn slash(
        &mut self, 
        params: &PosParams, 
        current_epoch: impl Into<Epoch>, 
        evidence_epoch: impl Into<Epoch>, 
        evidence_block_height: impl Into<u64>, 
        slash_type: SlashType, 
        validator: &Self::Address
    ) -> Result<(), SlashError<Self::Address>> { ... }

}

PoS system base trait for system initialization on genesis block, updating the validator on a new epoch and applying slashes.

Associated Types

type Address: 'static + Display + Debug + Clone + PartialEq<Self::Address> + Eq + PartialOrd<Self::Address> + Ord + Hash + BorshDeserialize + BorshSerialize

Address type

type TokenAmount: 'static + Display + Debug + Default + Clone + Copy + Add<Self::TokenAmount> + AddAssign<Self::TokenAmount> + Sub<Self::TokenAmount> + PartialOrd<Self::TokenAmount> + Into<u64> + From<u64> + Into<Self::TokenChange> + SubAssign<Self::TokenAmount> + BorshDeserialize + BorshSerialize

Token amount type

type TokenChange: 'static + Display + Debug + Default + Clone + Copy + PartialOrd<Self::TokenChange> + Add<Self::TokenChange> + Sub<Self::TokenChange> + From<Self::TokenAmount> + From<i128> + Into<i128> + Neg + BorshDeserialize + BorshSerialize

Token change type

type PublicKey: 'static + Debug + Clone + BorshDeserialize + BorshSerialize

Cryptographic public key type

Associated Constants

const POS_ADDRESS: Self::Address

Address of the PoS account

const POS_SLASH_POOL_ADDRESS: Self::Address

Address of the slash pool, into which slashed tokens are transferred.

Required methods

fn staking_token_address() -> Self::Address

Address of the staking token TODO: this should be const, but in the ledger address::xan is not a const fn

fn read_pos_params(&self) -> PosParams

Read PoS parameters.

fn read_validator_address_raw_hash(
&self,
raw_hash: impl AsRef<str>
) -> Option<Self::Address>

Read PoS raw hash of validator’s address.

fn read_validator_consensus_key(
&self,
key: &Self::Address
) -> Option<Epoched<Self::PublicKey, OffsetPipelineLen>>

Read PoS validator’s consensus key (used for signing block votes).

fn read_validator_state(
&self,
key: &Self::Address
) -> Option<Epoched<ValidatorState, OffsetPipelineLen>>

Read PoS validator’s state.

fn read_validator_total_deltas(
&self,
key: &Self::Address
) -> Option<EpochedDelta<Self::TokenChange, OffsetUnboundingLen>>

Read PoS validator’s total deltas of their bonds (validator self-bonds and delegations).

fn read_validator_voting_power(
&self,
key: &Self::Address
) -> Option<EpochedDelta<VotingPowerDelta, OffsetUnboundingLen>>

Read PoS validator’s voting power.

fn read_validator_slashes(&self, key: &Self::Address) -> Vec<Slash, Global>

Read PoS slashes applied to a validator.

fn read_validator_set(
&self
) -> Epoched<ValidatorSet<Self::Address>, OffsetUnboundingLen>

Read PoS validator set (active and inactive).

fn read_total_voting_power(
&self
) -> EpochedDelta<VotingPowerDelta, OffsetUnboundingLen>

Read PoS total voting power of all validators (active and inactive).

fn write_pos_params(&mut self, params: &PosParams)

Write PoS parameters.

fn write_validator_address_raw_hash(&mut self, address: &Self::Address)

Write PoS validator’s raw hash its address.

fn write_validator_staking_reward_address(
&mut self,
key: &Self::Address,
value: &Self::Address
)

Write PoS validator’s staking reward address, into which staking rewards will be credited.

fn write_validator_consensus_key(
&mut self,
key: &Self::Address,
value: &Epoched<Self::PublicKey, OffsetPipelineLen>
)

Write PoS validator’s consensus key (used for signing block votes).

fn write_validator_state(
&mut self,
key: &Self::Address,
value: &Epoched<ValidatorState, OffsetPipelineLen>
)

Write PoS validator’s state.

fn write_validator_total_deltas(
&mut self,
key: &Self::Address,
value: &EpochedDelta<Self::TokenChange, OffsetUnboundingLen>
)

Write PoS validator’s total deltas of their bonds (validator self-bonds and delegations).

fn write_validator_voting_power(
&mut self,
key: &Self::Address,
value: &EpochedDelta<VotingPowerDelta, OffsetUnboundingLen>
)

Write PoS validator’s voting power.

fn write_validator_slash(&mut self, validator: &Self::Address, value: Slash)

Write (append) PoS slash applied to a validator.

fn write_bond(
&mut self,
key: &BondId<Self::Address>,
value: &EpochedDelta<Bond<Self::TokenAmount>, OffsetPipelineLen>
)

Write PoS bond (validator self-bond or a delegation).

fn write_validator_set(
&mut self,
value: &Epoched<ValidatorSet<Self::Address>, OffsetUnboundingLen>
)

Write PoS validator set (active and inactive).

fn write_total_voting_power(
&mut self,
value: &EpochedDelta<VotingPowerDelta, OffsetUnboundingLen>
)

Read PoS total voting power of all validators (active and inactive).

fn init_staking_reward_account(
&mut self,
address: &Self::Address,
pk: &Self::PublicKey
)

Initialize staking reward account with the given public key.

fn credit_tokens(
&mut self,
token: &Self::Address,
target: &Self::Address,
amount: Self::TokenAmount
)

Credit tokens to the target account. This should only be used at genesis.

fn transfer(
&mut self,
token: &Self::Address,
amount: Self::TokenAmount,
src: &Self::Address,
dest: &Self::Address
)

Transfer tokens from the src to the dest.

Provided methods

fn init_genesis<'a>(
&mut self,
params: &'a PosParams,
validators: impl Iterator<Item = &'a GenesisValidator<Self::Address, Self::TokenAmount, Self::PublicKey>> + Clone + 'a,
current_epoch: impl Into<Epoch>
) -> Result<(), GenesisError>

Initialize the PoS system storage data in the genesis block for the given PoS parameters and initial validator set. The validators’ tokens will be put into self-bonds. The given PoS parameters are written with the PosBase::write_pos_params method.

fn validator_set_update(
&self,
current_epoch: impl Into<Epoch>,
f: impl FnMut(ValidatorSetUpdate<Self::PublicKey>)
)

Calls a closure on each validator update element.

fn slash(
&mut self,
params: &PosParams,
current_epoch: impl Into<Epoch>,
evidence_epoch: impl Into<Epoch>,
evidence_block_height: impl Into<u64>,
slash_type: SlashType,
validator: &Self::Address
) -> Result<(), SlashError<Self::Address>>

Apply a slash to a byzantine validator for the given evidence.