Module anoma_vm_env::tx_prelude::key::ed25519

Ed25519 keys and related functionality

Modules

testing
Helpers for testing with keys.

Structs

Keypair
An ed25519 keypair.

PublicKey
Ed25519 public key

PublicKeyHash
Ed25519 public key hash

SecretKey
Ed25519 secret key

Signature
Ed25519 signature

Signed
A generic signed data wrapper for Borsh encode-able data.

SignedTxData
This can be used to sign an arbitrary tx. The signature is produced and verified on the tx data concatenated with the tx code, however the tx code itself is not part of this structure.

Enums

ParsePublicKeyError

ParseSecretKeyError

VerifySigError

Functions

is_pk_key
Check if the given storage key is a public key. If it is, returns the owner.

pk_key
Obtain a storage key for user’s public key.

sign
Sign the data with a key.

sign_tx
Sign a transaction using SignedTxData.

verify_signature
Check that the public key matches the signature on the given data.

verify_signature_raw
Check that the public key matches the signature on the given raw data.

verify_tx_sig
Verify that the transaction has been signed by the secret key counterpart of the given public key.

Type Definitions

SignatureError
Errors which may occur while processing signatures and keypairs.