Trait anoma_vm_env::tx_prelude::key::SigScheme

pub trait SigScheme: Eq + Ord + Debug + Serialize + Default {
    type Signature: 'static + Signature;
    type PublicKey: 'static + PublicKey;
    type SecretKey: 'static + SecretKey;

    const TYPE: SchemeType;

    fn generate<R>(csprng: &mut R) -> Self::SecretKey
    where
        R: CryptoRng + RngCore;

    fn sign(
        keypair: &Self::SecretKey, 
        data: impl AsRef<[u8]>
    ) -> Self::Signature;

    fn verify_signature<T>(
        pk: &Self::PublicKey, 
        data: &T, 
        sig: &Self::Signature
    ) -> Result<(), VerifySigError>
    where
        T: BorshSerialize + BorshDeserialize;

    fn verify_signature_raw(
        pk: &Self::PublicKey, 
        data: &[u8], 
        sig: &Self::Signature
    ) -> Result<(), VerifySigError>;
}

Represents a digital signature scheme. More precisely this trait captures the concepts of public keys, private keys, and signatures as well as the algorithms over these concepts to generate keys, sign messages, and verify signatures.

Associated Types

type Signature: 'static + Signature

Represents the signature for this scheme

type PublicKey: 'static + PublicKey

Represents the public key for this scheme

type SecretKey: 'static + SecretKey

Represents the secret key for this scheme

Associated Constants

const TYPE: SchemeType

The scheme type of this implementation

Required methods

fn generate<R>(csprng: &mut R) -> Self::SecretKey where
R: CryptoRng + RngCore,

Generate a keypair.

fn sign(keypair: &Self::SecretKey, data: impl AsRef<[u8]>) -> Self::Signature

Sign the data with a key.

fn verify_signature<T>(
pk: &Self::PublicKey,
data: &T,
sig: &Self::Signature
) -> Result<(), VerifySigError> where
T: BorshSerialize + BorshDeserialize,

Check that the public key matches the signature on the given data.

fn verify_signature_raw(
pk: &Self::PublicKey,
data: &[u8],
sig: &Self::Signature
) -> Result<(), VerifySigError>

Check that the public key matches the signature on the given raw data.