Struct anoma_vm_env::tx_prelude::storage::Key

pub struct Key {
    pub segments: Vec<DbKeySeg, Global>,
}

A storage key is made of storage key segments DbKeySeg, separated by KEY_SEGMENT_SEPARATOR.

Fields
segments: Vec<DbKeySeg, Global>

The segments of the key in the original (left-to-right) order.

Implementations

impl Key

pub fn parse(string: impl AsRef<str>) -> Result<Key, Error>

Parses string and returns a key

pub fn push<T>(&self, other: &T) -> Result<Key, Error> where
T: KeySeg,

Returns a new key with segments of Self and the given segment

pub fn join(&self, other: &Key) -> Key

Returns a new key with segments of Self and the given key

pub fn find_addresses(&self) -> Vec<Address, Global>

Returns the addresses from the key segments

pub fn len(&self) -> usize

Returns the length

pub fn is_empty(&self) -> bool

Returns true if the key is empty

pub fn validity_predicate(addr: &Address) -> Key

Returns a key of the validity predicate of the given address Only this function can push “?” segment for validity predicate

pub fn is_validity_predicate(&self) -> Option<&Address>

Check if the given key is a key to a validity predicate. If it is, returns the address of the account.

pub fn is_updatable(&self) -> bool

Check if the given key can be updated

pub fn parse_db_key(db_key: &str) -> Result<Key, Error>

Returns a key from the given DB key path that has the height and the space type

pub fn sub_key(&self) -> Result<Key, Error>

Returns a sub key without the first segment