Struct anoma_vm_env::tx_prelude::token::Amount

pub struct Amount { /* fields omitted */ }

Amount in micro units. For different granularity another representation might be more appropriate.

Implementations

impl Amount

pub fn change(&self) -> i128

Get the amount as a Change

pub fn spend(&mut self, amount: &Amount)

Spend a given amount. Panics when given amount > self.micro amount.

pub fn receive(&mut self, amount: &Amount)

Receive a given amount. Panics on overflow.

pub const fn whole(amount: u64) -> Amount

Create a new amount from whole number of tokens

pub fn max() -> Amount

Create a new amount with the maximum value