Trait anoma_vm_env::imports::tx::BorshDeserialize

pub trait BorshDeserialize {
    pub fn deserialize(buf: &mut &[u8]) -> Result<Self, Error>;

    pub fn try_from_slice(v: &[u8]) -> Result<Self, Error> { ... }

    pub unsafe fn is_u8() -> bool { ... }
}

A data-structure that can be de-serialized from binary format by NBOR.

Required methods

pub fn deserialize(buf: &mut &[u8]) -> Result<Self, Error>

Deserializes this instance from a given slice of bytes. Updates the buffer to point at the remaining bytes.

Loading content...
Provided methods

pub fn try_from_slice(v: &[u8]) -> Result<Self, Error>

Deserialize this instance from a slice of bytes.

pub unsafe fn is_u8() -> bool

Whether Self is u8. NOTE: Vec<u8> is the most common use-case for serialization and deserialization, it’s worth handling it as a special case to improve performance. It’s a workaround for specific Vec<u8> implementation versus generic Vec<T> implementation. See https://github.com/rust-lang/rfcs/pull/1210 for details.

It’s marked unsafe, because if the type is not u8 it leads to UB. See related issues:

https://github.com/near/borsh-rs/issues/17

https://github.com/near/borsh-rs/issues/18

Loading content...