Module anoma_vm_env::tx_prelude

Re-exports

pub use crate::governance::tx as governance;

pub use crate::ibc::Ibc;

pub use crate::imports::tx::*;

pub use crate::intent::tx as intent;

pub use crate::nft::tx as nft;

pub use crate::proof_of_stake;

pub use crate::proof_of_stake::PoS;

pub use crate::token::tx as token;

Modules

address

Implements transparent addresses as described in Accounts Addresses.

chain

Chain related data types

dylib

Dynamic library helpers

governance

Files defyining the types used in governance.

hash

Types for working with 32 bytes hashes.

ibc

Types that are used in IBC.

intent

Intent data definitions and transaction and validity-predicate helpers.

internal

Shared internal types between the host env and guest (wasm).

key

Cryptographic keys

matchmaker

Matchmaker types

nft

Nft types

storage

governance storage

storage

Storage types

time

Types for dealing with time and durations.

token

A basic fungible token

transaction

Types that are used in transactions.

validity_predicate

Types that are used in validity predicates.

Structs

Signed

A generic signed data wrapper for Borsh encode-able data.

SignedTxData

This can be used to sign an arbitrary tx. The signature is produced and verified on the tx data concatenated with the tx code, however the tx code itself is not part of this structure.

Enums

Address

An account’s address

Traits

IbcActions

IBC trait to be implemented in integration that can read and write

PosRead

Read-only part of the PoS system

PosWrite

PoS system trait to be implemented in integration that can read and write PoS data.

Attribute Macros

transaction

Generate WASM binding for a transaction main entrypoint function.