Module anoma_vm_env::vp_prelude

Re-exports

pub use crate::imports::vp::*;

pub use crate::intent::vp as intent;

pub use crate::key::vp as key;

pub use crate::nft::vp as nft;

pub use crate::token::vp as token;

Modules

address

Implements transparent addresses as described in Accounts Addresses.

chain

Chain related data types

dylib

Dynamic library helpers

hash

Types for working with 32 bytes hashes.

ibc

Types that are used in IBC.

intent

Intent data definitions and transaction and validity-predicate helpers.

internal

Shared internal types between the host env and guest (wasm).

key

Cryptographic keys

matchmaker

Matchmaker types

nft

Nft types

parameters

Protocol parameters

proof_of_stake

Proof-of-Stake integration as a native validity predicate

storage

Storage types

time

Types for dealing with time and durations.

token

A basic fungible token

transaction

Types that are used in transactions.

validity_predicate

Types that are used in validity predicates.

Structs

BTreeSet

A set based on a B-Tree.

HashSet

A hash set implemented as a HashMap where the value is ().

Signed

A generic signed data wrapper for Borsh encode-able data.

SignedTxData

This can be used to sign an arbitrary tx. The signature is produced and verified on the tx data concatenated with the tx code, however the tx code itself is not part of this structure.

Enums

Address

An account’s address

Attribute Macros

validity_predicate

Generate WASM binding for validity predicate main entrypoint function.